refactored plugins library based on cryptosystem
