--- packages/plugins/src/crypto/index.ts
@@ -1,2 +1,0 @@
-export * as rsa from "./rsa.js"
-export * as ecdsa from "./ecdsa.js"
