--- packages/plugins/src/did/index.ts
@@ -1,3 +1,0 @@
-export * from "./prefix.js"
-export * from "./transformers.js"
-export * from "./validation.js"
