+++ packages/plugins/src/ed25519/crypto.ts
@@ -1,0 +1,10 @@
+import { EDWARDS_DID_PREFIX } from "../prefixes"
+import { didFromKeyBytes, keyBytesFromDid } from "../util"
+
+export const didToPublickey = (did: string): Uint8Array => {
+  return keyBytesFromDid(did, EDWARDS_DID_PREFIX)
+}
+
+export const publicKeyToDid = (pubkey: Uint8Array): string => {
+  return didFromKeyBytes(pubkey, EDWARDS_DID_PREFIX)
+}
