+++ packages/plugins/src/index.ts
@@ -1,5 +1,12 @@
-export * from "./keypair/ed25519.js"
-export * from "./keypair/rsa.js"
-export * from "./did/index.js"
-export * as keypair from "./keypair/index.js"
-export * from "./plugins.js"
+export * from "./ed25519/keypair.js"
+export * from "./rsa/keypair.js"
+export * from "./p256/keypair.js"
+
+export * from "./ed25519/plugin.js"
+export * from "./rsa/plugin.js"
+export * from "./p256/plugin.js"
+
+export * from "./default-plugins.js"
+
+export * as rsa from "./rsa/crypto.js"
+export * as p256 from "./p256/crypto.js"
