--- packages/plugins/src/keypair/index.ts
@@ -1,4 +1,0 @@
-export * from "./ed25519.js"
-export * from "./ecdsa.js"
-export * from "./rsa.js"
-export * from "./base.js"
