+++ packages/plugins/src/p256/keypair.ts
@@ -1,38 +1,42 @@
 import { webcrypto } from "one-webcrypto"
 import * as uint8arrays from "uint8arrays"
-import { Encodings } from '@ucans/core'
+import { Didable, Encodings, ExportableKey, Keypair } from "@ucans/core"
 
-import * as ecdsa from "../crypto/ecdsa.js"
+import * as crypto from "./crypto.js"
 import {
   AvailableCryptoKeyPair,
   isAvailableCryptoKeyPair,
   PrivateKeyJwk,
 } from "../types.js"
-import BaseKeypair from "./base.js"
 
-export class EcdsaKeypair extends BaseKeypair {
+
+export class EcdsaKeypair implements Keypair, Didable, ExportableKey {
+
+  private publicKey: Uint8Array
   private keypair: AvailableCryptoKeyPair
+  private exportable: boolean
 
   constructor(
     keypair: AvailableCryptoKeyPair,
     publicKey: Uint8Array,
     exportable: boolean
   ) {
-    super(publicKey, "p256", exportable)
     this.keypair = keypair
+    this.publicKey = publicKey
+    this.exportable = exportable
   }
 
   static async create(params?: {
     exportable?: boolean
   }): Promise<EcdsaKeypair> {
     const { exportable = false } = params || {}
-    const keypair = await ecdsa.generateKeypair()
+    const keypair = await crypto.generateKeypair()
 
     if (!isAvailableCryptoKeyPair(keypair)) {
       throw new Error(`Couldn't generate valid keypair`)
     }
 
-    const publicKey = await ecdsa.exportKey(keypair.publicKey)
+    const publicKey = await crypto.exportKey(keypair.publicKey)
     return new EcdsaKeypair(keypair, publicKey, exportable)
   }
 
@@ -42,18 +46,22 @@ export class EcdsaKeypair extends BaseKeypair {
       exportable?: boolean
     }): Promise<EcdsaKeypair> {
       const { exportable = false } = params || {}
-      const keypair = await ecdsa.importKeypairJwk(jwk, exportable)
+      const keypair = await crypto.importKeypairJwk(jwk, exportable)
 
       if (!isAvailableCryptoKeyPair(keypair)) {
         throw new Error(`Couldn't generate valid keypair`)
       }
 
-    const publicKey = await ecdsa.exportKey(keypair.publicKey)
+    const publicKey = await crypto.exportKey(keypair.publicKey)
     return new EcdsaKeypair(keypair, publicKey, exportable)
-    }
+  }
+
+  did(): string {
+    return crypto.publicKeyToDid(this.publicKey)
+  }
 
   async sign(msg: Uint8Array): Promise<Uint8Array> {
-    return await ecdsa.sign(msg, this.keypair.privateKey)
+    return await crypto.sign(msg, this.keypair.privateKey)
   }
 
   async export(format: Encodings = "base64pad"): Promise<string> {
